Runtime fixes, DOCX save modernization, Windows workflow, CI for clang x64/ARM64
  Fixes command dispatch (exports .def), KME pointer tables, PropToRtf,
  and StringMap embedded-NUL truncation. Modern-format saves adopt the
  chosen name and clear the dirty flag (project decision). Adds windows/
  helpers for the git-less build machine and CI jobs for the proven
  clang-x64 path plus a native ARM64 build+test on windows-11-arm.

+++ src/Opus/spelcore.c
@@ -154,7 +154,7 @@ int TmcGosubSpellMM()
 	pcabspellerMM->iSplMMAddToBox = vrf.udcDefault;
 	pcabspellerMM->fSplMMAutoSugg = vpref.fSplAutoSugg;
 	pcabspellerMM->fSplMMIgnoreCaps = vrf.fSplIgnoreCaps;
-	(int)(pcabspellerMM->uSplMMSuggList) = uNinchList;
+	pcabspellerMM->uSplMMSuggList = uNinchList;
 
 	cmb.hcab = hcabspellerMM;
 	cmb.cmm = cmmNormal;
